Rename qos_event.hpp -> event_handler.hpp
